Fix invoice/receipt business name, fractional stock, sale channels
Three things, discovered to be more connected than they first looked.

Root cause of the invoice/receipt heading problem: invoice_print.html
and payment_receipt.html already correctly render
{{ invoice.tenant.name }} -- the template was never the bug. TenantModel
.save() was: it had a stub with a comment reading "Set tenant from
request if not set... But better to set in views", followed by a bare
`pass`. Nothing ever finished it. quick_sale's Invoice.objects.create()
(and InvoiceItem, and Payment) never passed tenant= explicitly, so every
quick sale invoice silently saved with tenant=None -- invisible until
someone printed one and saw "Business" instead of their real business
name. Fixed at the base class: TenantModel.save() now pulls the current
tenant from the thread-local get_current_tenant() when tenant isn't
already set. This fixes every model that inherits it, not just Invoice --
confirmed by creating an Invoice exactly the way quick_sale does, with
and without the fix.

Fractional quantities: Product.quantity and minimum_stock were
IntegerField, so a business selling by weight or volume (2.5 kg of rice,
1.5 litres of oil) couldn't stock or sell fractional amounts. Both are
now DecimalField, matching InvoiceItem.qty and OrderItem.qty, which
needed the same change -- stocking 2.5kg is pointless if a sale can only
move whole units of it.

Wholesale/online sale recording: Product already had retail_price,
wholesale_price, and online_price, but quick_sale always used
retail_price regardless, and nothing recorded which tier a given sale
actually used. Added InvoiceItem.sale_channel (retail/wholesale/online,
default retail) and a channel selector per line in Quick Sale; price
auto-fills from the selected channel's price on the product (client-side
via data-retail-price/data-wholesale-price/data-online-price attributes)
but is still resolved server-side too when the client leaves price
blank, so the recorded price can't be spoofed by tampering with the
channel dropdown.

Also fixed while touching the same function: product_import_csv passed
business=request.user.business if hasattr(...) else None to
Product.objects.create() -- Product has no business field (only tenant,
inherited from TenantModel) and User has no .business attribute either,
so hasattr() was always False and the call always threw
TypeError: invalid keyword argument 'business'. Caught by a broad
try/except and reported as a per-row error, every CSV import row was
silently failing. Verified fixed: a CSV row with a fractional quantity
now imports correctly with the tenant properly assigned.

All verified through the real view (Django test client, not just model
saves), each check rolled back afterward: quick_sale with a wholesale
line resolves to wholesale_price server-side (not retail_price), records
qty=2.50 and sale_channel=wholesale, decrements stock by exactly 2.5,
and the resulting invoice has a real tenant instead of None. CSV import
separately verified the same way.

+++ inventory/models.py
@@ -22,8 +22,10 @@ class Product(TenantModel):
     wholesale_price = models.DecimalField(max_digits=10, decimal_places=2)
     online_price = models.DecimalField(max_digits=10, decimal_places=2)
     
-    quantity = models.IntegerField(default=0)
-    minimum_stock = models.IntegerField(default=5)
+    # Decimal, not integer -- goods sold by weight/volume (e.g. 2.5 kg of
+    # rice, 1.5 litres of cooking oil) need fractional stock levels.
+    quantity = models.DecimalField(max_digits=10, decimal_places=2, default=0)
+    minimum_stock = models.DecimalField(max_digits=10, decimal_places=2, default=5)
     
     supplier = models.ForeignKey(Supplier, on_delete=models.SET_NULL,
                                  blank=True, null=True)
